refactor(api-languages): use @core prisma alias and share WESS import helpers
